🖥️ ui(web): introduce ChallengeInfoView component to enhance challenge information display and refactor ChallengeInfo for improved structure
